ACDC-GAN/requirements.txt
matplotlib~=3.3.4
numpy~=1.19.2
pillow~=8.3.1
tqdm~=4.63.0
keras~=2.3.1
scipy~=1.2.1
imageio~=2.9.0
tensorflow-gpu=1.14